Added support for download programs that forward the download action to a remote computer.
This way a download manager like aria2 running on some server somewhere on the
network can be directly used by MediathekView to initiate downloads of films.

A Linux bash script was added as an example program to use with this new feature.
It forwards the download request to an aria2 daemon (version 1.19 and above)
secured with a secret token. The program parameters for this bash script must be:

%f ** <URL of aria2 daemon> <secret for aria2 daemon>

The URL is the full URL of the daemon including the port number and path
(i.e. http://somewhere.com:6800/jsonrpc).

The settings page of a program of a program set now has an additional check box
to mark a program as such a remote downloading program. Also, due to the fact
that no local file will be created anymore, the download status and download
start button of the download list had to be modified to act accordingly.


Former-commit-id: 8899c8dd4d64556c4d4630f0d51436e4f450a8ef